chore(build): remove incompatible pointer suppression
Keep PsyQ-shaped CD/SPU API boundaries and make byte-buffer crossings explicit
at the call sites.

Constrain the native decoded-SPU clear for CD-only reads and assert multitap
packet layout so the broad incompatible-pointer suppression can be removed.

+++ game/HOWL/HOWL_Bank.c
@@ -187,7 +187,7 @@ int Bank_AssignSpuAddrs()
 			// start transfer
 			SpuSetTransferStartAddr(spuAddrStart);
 
-			SpuWrite((u32 *)((int)sdata->ptrSampleBlock2 + 0x800), (size_t)sdata->audioAllocSize);
+			SpuWrite((u8 *)((int)sdata->ptrSampleBlock2 + 0x800), (size_t)sdata->audioAllocSize);
 		}
 
 		sdata->bankLoadStage++;
